+++ pages/markdown/misc/changelog.md
@@ -1,5 +1,11 @@
 -----
 
+## Deck Updating - Jan 03 2020 - 0.3.1
+This small update fixes some undetected issues in pulling deck changes.  You can now update the decks that you copied from shared decks properly.
+
+
+* **Fixed Flashcard Pulling:**  Fixed a major issue that would cause pulled flashcards to appear in your deck, but not be able to be studied.
+
 
 # Games! - Jan 02 2020 - 0.3.0
 Flashcards don't just have to be work; they can be fun too!